+++ Tests/PrimitiveTests/IntMacroTests.swift
@@ -1,0 +1,16 @@
+import XCTest
+@testable import RAW
+
+
+final class MyTests:XCTestCase {
+
+	func testArray() throws {
+
+		let buildInt = My64BitThinggy(0x1234567890ABCDEF)
+		// Int.KMS()
+		// let asInt = My64BitThinggy(buildInt)
+
+	}
+
+	
+}